got cds query working for flybase. fb doesnt use cds, its figured from the protein, so i wrote a whole different query as it was rather different. Not sure if this is the best way to go - probably should revisit this - but it gets it working for now.
cds (proteins) are
showing up as separate features - need to get rid of this for fb - fb
just shows starts and stops in the transcript (also the cds's im
getting span introns - i imagine thats not the case for tigr)

also the gene query and cds query need to be optimized - both are
taking a minute - much too long.

also sequence is being retrieved - probably could put that off and
retrieve it on demand.