Added gene query for flybase - this was refreshingly easy to do. So to get synonyms  - like common gene names like "cact" - JdbcChadoAdapter.getFeatureId is querying feature for both uniquename and name. The gene 'cact' has uniquename CG5848 and name 'cact'. I doubt that it will but this could bring back more than one feature - i dont think it will in that i doubt theres features whose name matches a separte feature uniquename - but if this is an issue this can be changed to 1st query uniquename, and if that gets nothing then query name (or vice versa).
As has been mentioned seq types that get queried on need to be
configured - it would be nice to configure these in style file, except
that you need these at data adapter gui time which may be awkward -
but i suppose DatabaseToStyle could be used - need to think about this.
in general it might be nice to migrate chado-adapter.cfg stuff to
either a style or apollo.cfg file.