Got DatabaseToStyle working for chado adapter. So now loading game.style for flybase chado. Problem was setDatabase has to happen before setStyle. so i moved setDatabase to CAGUI.doOp, before call to getCuration.
Uses the database field from chado adapter, which seems like the
proper thing, except this field refers to the actual database name on
the database server, and that name might not be unique across servers
- should probably change this to use database label which is unique(or
it would be silly if it wasnt).

Need to get flybase chado spitting out types from game.tiers.