gene names are now coming up proper in fb chado - if they have real name (eg cact) that is now its label, instead of the CG id. This is a divergence from tigr. tigr uses chado feature uniquename for both the apollo features name and id, flybase now uses uniquename for the id (CG#) and feature.name for the features name.
I finally put some configging in this - as a simple start (we can get
more involved as we need to) i just added

ChadoInstance  TIGR | Flybase

to the style file. So for now all differences will just be denoted
with this one flag - this will be expanded as we go.

So this led to making a ChadoInstance interface and a
FlybaseChadoInstance and TigrSybilChadoInstance classes implementing
it. I added this new gene name issue to it and migrated a few things -
but theres still a few things to migrate over from potgres/sybase
classes, which have non-postgres-sybase stuff in them. also the cv
stuff should probably be broken out as a class, and featureloc classes
should probably be outer classes as well, but i made some progress,
didnt wanna go all the way down the slippery slope at once.