harvard reported a problem reading some XML files.  it turned out they included annotations of type "misc. non-coding RNA", which was not defined in their game.tiers, and therefore had no name adapter defined.
i figured i could fix that by making sure that when a new feature
property is created (as it is for "misc. non-coding RNA"), it gets
assigned the name adapter that was installed in the style.

the problem is that later on, when the name adapter is requested, it
comes back null, because that's how we set up getNameAdapter:

  /** Return null if no tiername or tiername is not "annotation" */
  public ApolloNameAdapterI getNameAdapter() {

when "misc. non-coding RNA" is read in the data, it's added as a new
tier, which is not called "annotation".

or is it really crucial for getNameAdapter() to return null if the
tiername is not "Annotation"?  (that seems a little dangerous to have
hardcoded, actually.)  making it just return the name adapter, regardless
of tier name, fixes this problem and doesn't have any obvious bad
effects.