Fixed 2 synteny bugs:
Previously I was using

Transcript.getPeptideSequence().getAccessionNo();

to get the transcripts peptide name (eg CG3415-PA). Well that method
now returns null. And the peptide name is now found in the display
id. Not sure why this changed. So im using the display id. Should probably add a
"getPeptideName()" method or something like that to make this explicit
- it shouldnt be linking to a display name.

Also put in check that history # of species is not more than the
configged max - this can happen if config changes - was causing exception.