1. in order to get the current highest pixel for a given tier the transformer needed to be passed as an argument when new features are added so that this can be calculated (fixed the expand problem)
2. fixed all the naming isses and along the way consolidated the
naming code in SetDetailPanel and in NameAdapters so that both
tables are now using the same methods. As a minor issues the game
adapter no longer sets the db for an analysis if the value of
the db is "dummy"

The biggest problem causing the bogus names was in the GAME
adapter. Someone was creating FeaturePairs for gene prediction
results. Doh! There was even a comment saying that if there
is a seq_relationship element, then there must be a pair.
No way. There is only a pair if there is a seq_relationship
to something other than the curated sequence. Don't forget this!!!!