Fixed bug where expand/collapse tiers was not working (due to my rewrite). Problem was the TypesObserver was set up with the master controller, so the TypesChangedEvents were not getting to the curation controllers.
So in fixing this i did a bit of
refactoring. FeatureProperty, TierProperty, PropertyScheme, and
TypesObserver were all using Observer/Observable to notify that the
prop scheme has changed. TypesObserver when notified would send out a
TypesChangedEvent to the controller (the master controller). So the
Observer/Observable pattern is a nice pattern to use, but in this case
its redundant with TypesChanged event/listener. Event-listener is
basically a more involved Observer-Observable. Its confusing to have
both. (I imagine theres some history here - like the Observer stuff
pre-dates the events).

So I took out Observer-Observable and not PropertyScheme is firing
events, and TierProperty and FeatureProperty tell PropertyScheme to
fire events as well. TypesObserver is no longer relevant and has been
removed. I renamed TypesChangedEvent/Listener to
PropSchemeChangeEvent/Listener. These events only say that the
PropertyScheme has changed they do not give any specifics. TypePanel
and FeatureTierManager when they get the event just totally redo
themselves, so theres actually no need to be more specific about the
changes.

I also moved these renamed events to apollo.config as they are really
config events. PropertyScheme is in config and produces these events,
so the events dont belong in gui, they are "bigger" than just gui. I
feel this is true of other events in apollo.gui.event, like
RegionChangeEvent, that also need to migrate to other packages, as a
lot of events go beyond strict gui-ness.

So now CurationState provides PropertySchemes, as it should, and sets
the property schemes PropSchemeChangeListener to the curation
controller. Presently the way synteny works is that there is only one style and
tiers/propScheme for all curations. I think eventually we are going to
want synteny to be more sophisticated and be able to have different
property schemes for each curation - like if curations came from
totally different sources that have different types and all. The
changes here are a step in that direction, and when we go that way it
will now be easier.