allen as scott again ;)
this fixes the broken search.  the problem was that a workaround to emulate
apache1 behaviour from apache2 was actually breaking the apache1 behaviour.

we now check to see if we are running under apache1, and if so do the old
apache1 thing.

is that vague enough w/o reading the diff?