Fixed AnnotationChangeLog for multi-curations. This is basically the tail end of the rewrite - and support for undo and synteny. Theres a new editor package with AnnotationChangeLog, AnnotationChangeCoalescer, AnnotationChangeEvent, ResultChangeEvent, Transaction...
Also I added 2 new value classes TransactionOperation
and TransactionClass. Transaction uses them - eventually
FeatureChangeEvents should use them - I sort of cut this whole part
short as I want to get on to the suddenly urgent writeback task.

So there is now one AnnotationChangeLog per curation - which makes
dealing with transactions much easier.

The name adapters need to be moved to editor. didnt do this now for
expediency.

Name Adapter creation is different. a name adapter depends on
annotation change log as it needs to see which ids have been made when
making a new id. annotation change logs are particular to a curation,
so name adapters are too, but they are created by FeatureProperty
which can be multi-curation. So CurationState asks FP to create name
adapter, then it gives the new NA its annot change log, and CurState
holds on to it, rather than FP.

CurationState got split into dataadapter.CurationState and
gui.synteny.GuiCurationState as there needed to be CurationState in
the data world that was seperate from the gui. I contemplated having
it be an ISA but that didnt work as some non-gui cur state needs to be
set up before the gui comes into play, so ya need to initialze a
non-gui cur state and then have it associated with GuiCurState at gui
time. For now the read data adapters carry the data cur state for gui
cur state to grab and that seems to work ok.