Fixed bug I discovered while demoing apollo to guanming, which was from my recent changes. When type was changed in annot info editor an exception would fly on commit. This was because of my change where all AnnotChangeEvents from FeatureEditorDialog are fired at the controller, and AnnotationChangeLog gets them from listening for the events - previously were sent straight to ACL. Problem was type change was causing a REPLACE ANNOTATION event, so DrawableFeatureSet would replace the annotation. But then FED fires another REPLACE ANNOTATION, as it always has on commits, and it blows up as the replace has already been done.
So I changed REPLACE ANNOTATION for type change to
UPDATE TYPE. This brings up issues that are very topical that still
need to be worked out in a rigorous fashion. TYPE is really a subpart
of an Object not a real ObjectClass itself - UPDATE says just change
this small part of me not all of me. I think the way this is going is
toward another value class that describes subparts of objects so
updates can be fully described, but we are not there yet.