so i implemented a very primitive undo on the dicistronic boolean property and it actually works. i was by no means intending to do undo but the FED has a primitve undo already, which was gonna be hard to reimplement with fine gran transacs - so putting in a primitive transaction undo for the fed has proved to be the same amount of work if not easier. Im going to try it out on other parts of the fed but it seems a promising way to go - rather than reengineering fed commit to basically do a "redo" with all the events produced in the session that would have to be tracked, yada yada...
i should mention this is the low hanging fruit of undo - dicistronic
is a boolean - easiest case - and none of the fed events/trans are
compound like the main window (where moving exons results in a zillion
events) - presently i have undo so it only undoes stuff from the FED -
the rest is ignored

The transaction manager is coalescing these dicistronic events - which
it needs to do for the db so its nice to see it working - but for undo
we probably want them uncoalesced - this is nothing pressing now - it
can at least undo the last event which is pretty novel - but
eventually we will either want a separate trans manager for undo or do
coalescing at the time of write. the latter may be easier because if
an undo happens on an event that has been coalesced how does the db
tran man deal with that?