So GMOD-Web seems to have been broken for a little while. I fixed some of the problems in Makefile.PL. After installing on a clean box it became obvious that Makefile.PL was missing some required prereqs.  But the major reason gmod-web is currently not working is directly a result of the capitilization change that happened a while back.  When I updated Turnkey to support multiple Turnkey installs on the same box I changed the capitilization of the generated TT2 files (lowercase to first letter capitilization). I don't recall why I did this but it ended up breaking gmod-web.  The macros.tt2 file (in tt2/custom) expects all the tt2 files be all lowercase. It therefore doesn't include the proper files and distplays only the header.
Anyway, I know how to fix this problem but I'm actually going to work on Turnkey a bit and get that codebase
stable before I commit more changes here.  That way I can just push the changes from the updated Turnkey to
gmod-web and make any gmod-web bug fixes in the same go.  Depending on how long it takes to stabilize Turnkey
gmod-web may be working by the end of the week.

--Brian