silenced '... not numeric ...' errors w/ additional typecheck.
updated xml:base in several templates to give the actual name/version/etc of
the datasource rather than a string copied right out of the spec.

modified ControllerI to use my 'chado1' database.  hardcoded for now -- this
really needs to be moved to a config file or PerlSetVar statements (maybe one
of the latter pointing to one of the former).

region command now works on chado adaptor.  hardcoded to only look for features
of type 'chromosome' as regions.  ongoing discussion on gmod-schema list as to
how to make a reference sequence query type-agnostic.