AnnotationEditor.setTranslationTerminus now sending out events as TRANSLATION_RANGE not start and stop - this officialy makes translation start and stop irrelevant. just calls ACC.execChanges which checks for translation changes - so doesnt need to fire them itself.
translation range changes were going undetected due to bug in
FeatureList - if feature list was wrapping a vector a addAll would not
add all of its features. took out the whole vector wrapping thing -
now just dumps the contents of the vector into the list - much simpler
and not a big deal.