Did a bunch of work with splitTranscript. The events are now firing in the proper order - with the add new gene happening before the updating of the exon to that new gene - this was necasary for writing back to the db. Also there was somewhat redundant split stuff that was causing a bug. If there was just one transcript in the gene and it was split it would retain the gene name for one of the split genes. According to lynnes email with sima all splits should end up with 2 new temp ids and no retaining of old id - not even for the 1 transcript case. There was logic that covered some of the split cases in splitTranscript that did the naming wrong and was not needed anymore since splitAllGenes covered the one transcript case as well as the uncovered multi trans case.
purgeTranscript now just adds delete trans event and doesnt excute the
coalescer. AnnotationChangeCoalescer should be renamed ChangeList -
thats really what it is at this point - it doesnt do colescing - thats
done in the TransactionManager. in general subroutines should not be
firing events but just adding them to the change list for the caller
to do the firing - especially if the caller has to correct the
ordering.

need an event for changing the name to temp name - the chado adapter
can then decide whether that will be delete & add or an update of
id. chado trigger currently doesnt handle id update unfortunately.