it turns out that RELAX-NG (which can be stored as an rng file, which is XML, or an rnc file, "compact" format) is a better way to represent the GAME schema than XSD is.  rng can represent the idea that there are elements that can occur in any order, but that some should only occur once (e.g. one "start").  xsd allows elements to be in any order zero or more times--it doesn't allow us to say that they should occur in any order but zero or one time.  (you can, in xsd, say that elements should occur zero or one time *in a particular order*, but we don't want to impose an order constraint on GAME.)
the good news, though, is that we can convert the rng to an rnc and the
rnc to an xsd file that is valid but not as stringent-it allows any
number of "start" elements.  so if an xml file validates with the xsd,
that doesn't mean it will validate with the rng (which is more
restrictive).

note that any further schema editing should be done on the rng file,
since we can auto-generate a reasonable xsd from the rng but not
vice-versa.  unfortunately, when you convert from rnc to xsd, the
comments are lost.