split transcript may or may not incur a splitting of the gene. in the case where it doesnt it just does a delete and add transcript. in the case where it does split the gene the adding of the transcript is subsomed by the adding of the gene - but the split gene method causes a delete transcript - the transcript has to be added for the split gene to do its analysis. theres 2 possible solutions here - leave in the delete transcript and make sure to put in the add transcript before it - or just take out the deleting of the transcript. I went with the latter rather than having a spurious adding and deleting of the transcript.
I put isIdentical in RangeI - this could be renamed equals at some point
if need be. this will return true for a range and its clone. the
editor has to clone deleted trans for events before they get added to
new genes, so this method was needed to figure if clone and trans were
same.

i noticed the name and id events are still not quite right - need to
fix that. guanming, thats coming next.