fixed a annot editor merge event bug. on merging if 2nd gene name was chosen annot editor actually uses the 1st gene and renames (i think this is historical). the id change was happening after the name change which was messing things up - put it after.
but the transcript id change actually was causing a problem too as it
was being made into a add and delete of a gene (not a transcript) with
the transcript name - this aint right. ive supressed this event for
now - looking further into it. i think we need this transcript id
event - unless it happens automatically when the gene id changes -
does it? if not we need this.