committing new rice.tiers that has commented out the insertion result types:
A) they are a blind copy of the drosophila types with drosopila suffix
(.dros) in the resulttype - clearly useless to rice

B) the insertion site annot types do not yet exist.

for now taking this out. we can add back when all this is worked out
if we decide we want insertions in rice.