Made database lookup error more informative in 2 ways. When a lookup fails it returns -1. The old message would say id -1 is not found, now it prints the query and states that nothing was returned from the query.
Failed lookup would cause an apollo data adapter exception which org
bdgp DataAdapterChooser would show in a popup with
DataAdapterException.toString(). toString was rather uninformative as
it didnt tell much about the parent exception that data adapter
exception was wrapping, which is really what is of
interest. DataAdapterException.toString now calls parent exception
toString if there is a parent exception.

I think apollos DataAdapterException should be renamed
ApollAdapterException as it is exceptionally confusing that theres
both org bdgp DataAdapterException and an apollo DataAdapterException.