Fixed split commit bug. Splitting a gene with more than one transcript where a new transcript was added before the split turns out to be a funny case. So a new transcript is added first and is in the transaction list 1st. then a split comes along and splits the gene in 2 creating 2 new temp genes. now when this gets saved to the database even though the new transcript is being added to what now is a new gene it doesnt get coalesced out because the transaction record doesnt record this new gene happening until after the new transcript event. its a rather sneeky case.
but i found what i think is a simple solution to it. if the
annot/transcript to be added has a parent with a temp id, then
coalesce it out. if it has a parent with a temp id then the parent in
theory should get added in at some point (and if it doesnt then
something is awry) and the transcript/child will get committed to the
db when the parent does. so i just check for a temp id in the parent.

at the moment im just looking for "temp" in the name - at some point
this should be relegated to the name adapter.

i was actually doing this generically for all annots, but for gene
(root annots) the gene parent has name NO_NAME but
FeatureSet.getName() will return the 1st childs name if it is NO_NAME
which was problematic in this case. so i check for rootAnnotation() first.