fixed bug with committing splits that had previous edits to them. i fixed this for adds but it turns out the same issue comes up for updates and deletes - and i didnt like my fix for adds. that is when a split happens a chunk of gene that becomes a new gene - if that old chunk has old edits on it they were not getting coalesced out because the add was ahead of instaed of behind the edit - coalsecing was only happening on edits ahead of the add. splits turn that logic upside down. my previous thing felt sort of hacky of just looking for temp ids. now a feature (non-subpart) add looks at previous transactions to see if any are actually descendant from it and coalesces them.
in the future i think what should happen is that updating of parents
and name changes and type changes should be tracked. this will work
for now - but with an integrated db and once we have the ability to
request new temp ids from the db (cant do this yet without doing
insert itself) we will probably do a split as a name change and
updating of transcript parent and all - and the coalescing would track
this and use this info.

>