got undo of gene name working - its a bit of a kluge at the moment - it needs a revisit - so editing gene name also causes transcript name to change and theirs transactions for both - this is a first having to do undo for essentially a compound event/transaction edit. so for now i just undo all the name changes on the top of the transaction stack.
in the future i can see 2 ways of going about this. the transcript
name transactions could be scrapped and be implied by the gene name
transactions - although this may vary for different name adapters so
perhaps thats a little presumptious - could query the name adapter i
suppose. or this could also be done with a compound transaction - we
are certainly going to need them at some point.

exon names are not affected by gene name changes as they use the gene
id. currently rice doesnt have names for proteins so rice proteins are
unaffected by gene name change (unless this changes) - fly on the
other hand does have protein names based on gene names and this will
need to be ammended for fly.

also i fiddled with a new architecture for undo where the undo is
actually done by the Transaction itself - like the Command design
pattern i suppose - and i also experimented with making a NameSubpart
subclass that does the actual modifying of the model so transaction
can be fairly generic.