well its no coincedence im doing type undo last - it is definitely the hardest undo in annot info - because it changes ids and names as well. (it would be nice if ids were type independent). im still working on this but im at a commit point - type is undoing and ids are undoing, names are not undoing yet - shouldnt be too hard to get names.
well this took a while but id actually like to revisit this at some
point. so at the moment when a type is changed theres a type change
event, and if it caused ids and names to change there are no
events/transactions for that. the reason is type changes go through as
delete and add - since different types have different amount of levels
its hard to do it with updates - also what does it mean to update to a
type with no transcripts when you have integrated data for that
transcript(unclear how that should be handled - tricky issues here!).
also id updating is presently treated with a delete and
add as well.

the problem then of not having the id and name transactions is that
you then have to derive the id changes oneself on undo - as you dont
have the transaction record of what happened. and to change ids one
needs the name adapter as its species specific. so for now im passing
the name adapter in.

eventually it would be nice if all the id and name events were spit
out. also id event shouldnt be split into add and delete until
coalesce/commit time, not at add trans time.

in the interest of simplifying and only editing the model in one
place, it makes sense actually for the transactions to edit the model,
both for undo and do. im now conceiving of annotation change events as
just carriers of transactions - ACEs and Trans have always been rather
redundant. so an editor can create a transaction and tell the
transaction to edit the model - and fire the transaction via an
Annotation change event. annotation change events could even be
renamed transaction events. then on undo the transaction manager takes
the transaction off the stack and tells it to undo itself. this is
consistent with the Command design pattern.

fixed bug in rice name adapter - exon id was using annot name instead
of annot id.

need to do name undo, also theres a bug with underscores in rice id names.

(commented out compile errors from nomi)