So I fiddled around some more with game saving and managed to speed things up even more. so for a 76M game file - AE003628.4 - (which is a ridiculous size - the data should be pruned somehow) - buffered save was still taking a while - sequences were taking 68 seconds and analyses were taking 86 seconds - 2.5 minutes all together! I managed to get sequence saving down to 3 seconds and analyses down to 13 seconds - from 2.5 minutes to 16 seconds! woohoo!
the trick was using the buffered printwriter instead of building up and
passing around strings. GAMESave.writeSequences was building up a HUGE
string and passing it back to writeXML to write it to the (new)
buffered PrintWriter. now i pass the printWriter into
writeSequencesToWriter that writes the sequences directly to the
buffered writer. this turns out to be far quicker.

nomi - the chado xml writer should employ this technique in
general. ill go ahead and make some changes there. future writing
(writing you are about to implement) should also write to the
buffered writer and not a stringBuffer/String.