put game writing optimization for results in chado xml writer - so it now writes each result to the buffered writer rather than a string buffer.
writeSequences is not implemented yet - nomi you defintiely want to do
this trick for write sequences as well. i didnt put it in for genome
position, begin, end or annotations as they are smaller strings - you
could for consistency. also the passing of the print writer and
writing to it could be taken deeper - the deeper the better, the less
strings and the more writing to the writer the better. certainly any
new code you add you should use the writer instead of a string buffer
- no reason to use a string buffer when theres buffering in the
writer.