added ability to configure gene prediction programs from chado-adapter.xml:
    <chadoInstance id="riceInstance" default="true">
        <genePredictionPrograms>
            <program>FgenesH</program>
        </genePredictionPrograms>

can add more than one program. Scott is configuring up a a new chado
database so this ability is pretty important so ya he doesnt have to
write java code. need to do the same with alignment programs - but im
out of time - when i get back. im sure their will be more that needs
to come out in the config file out of the Chado instances.