This is an interim commit for undoing id changes. Presently it undoes the gene id, but the transcript ids dont yet undo. also if the id change caused a name change the name does not undo yet (actually it undos if you hit undo a second time which is wrong - it has to be atomic).
an id edit is a compoound edit - as names and transcript ids
potentially get changes (according to name adapter). ive been hacking
around compound transactions - but now im seeing its easier to just
implement them already than to hack around them.

so for id undo for flybase im going to create the id transaction in
the GEP and have the transaction edit the model - and it will be
compound! - i think this is the way to go - and is actually probably
easier than the hacks around it - transaction manager will not make a
delete and insert transaction for the ACE because it will already have
the transaction in tow - this contrasts with merges from AE that cause
ID events that get turned into del & add in TM - cant have del & add
for undo - no way to know thats an id undo - del & add was to hack
around the fact that rice/chado didnt have way to update to a temp id,
basically request a new id - i think it does now and this should be
ammended - but for GeneEditPanel id changes it never gets updated to a
temp id (or shouldnt at least) so dont have to worry about that issue
- also rice doesnt allow id editing (its disabled) so we are ok on 2
accounts - for integrated db del & add will have to be ammended to
true id updates to temps for merges - since rice is not integrated
(and no one else is using it) this is not pressing - but thats the
state of things.