Did some reorganizing of the way GAME adapter goes through the XML parse tree and fills in the Apollo datamodels.  It still starts with the entire parse tree (in order to save memory, we really should handle each element as it's read, but this is more reorganizing than I was prepared to do at the moment) but now (thanks to a new method in org.bdgp.xml.XMLElement) it pops off each child of the tree and deals with it.  Theoretically, this should reduce the maximum amount of memory used, although in practice I haven't observed much of a difference.
One win of this new approach is that I can make the progress bar slowly
move as elements are processed, which gives the user the comforting impression
that things are happening (which, indeed, they are).

Note that you need the new org.bdgp.jar (which I also checked in).


I also consolidated the annot and result processing so we don't go through
the list twice (once looking for annots and once looking for results),
though we still go through it to get seqs.

I reindented the whole buffer because a lot of the indenting was funky.