changed the structure of name adapters - used to be:
DefaultNameAdapter -> Fly -> Rice

now its

Default -> GmodNameAdapter -> Fly
                           -> Rice

so in stead of rice overriding fly functionality rice and fly are
overriding default gmod functionality. this is a work in
progress. presently gmod has a lot of fly stuff in it as it is still
evolving out of fly and into gmod/rice. eventually gmod will become
less flyey and more gmody/ricey.