fixed funny issue where one could create genes with many overlapping and identical exons. this was true either collapsed or not collapsed - if not collapsed could still select several results with overlapping ranges and drag them in (perhaps the gui should disallow dragging of overlapping results regardless of collapsedness?). now buildTranscript checks for overlapping exons and merges them. other ways and methods did this checking (move exon to trans) but buildTranscript never did.
while i was mucking with annotation editor i took the oppurtunity to
refactor it a bit - mergeExons now makes a CompoundTransaction -
eventually everything will go compound or at least transaction - this
is the next big step for undo - so i did a little of it with these
changes - why not - purgeExon & purgeGene & purgeTranscript do the same as they
are called by mergeExons. im wondering if ChangeList should become a
transaction list - and maybe that already covered by CompoundTransaction?