so i hadnt done much with peptides with rice/gmod as i didnt think it mattered. in fact changeSeqNames in gmod adapter was no-oped - as i thought apollo didnt need to care as the chado triggers would take care of peptide naming - which they do, and peptide names are never displayed in apollo. but without an explicit name it was getting the default transcript name - turns out that was fouling the chado commit as it was adding exons to proteins instead of transcript as it looks for features by name/id and theres a name ambiguity with the default.
so this bug forced me to work out peptide naming for gmod. this is not
fully done yet. so i changed the name adapter to not call the low-down
Transcript.derivePeptideName, and replaced it with the new
generatePeptideName (in name adapter)
which is generalized for rice and fly with
getTranscriptSuffixRoot and getPeptideSuffixRoot and
transcriptOrdinalIsNumeric. so presently gene & transcript name
changes in annot info will assign the proper unique peptide
names. this is not yet so for new annots & transcripts which still
rely on Trans.derivePeptideName - thats next and a bit stickier.