Transcript.get_cDNA was nulling out peptide, as a way getting the peptide to recalculate. the problem was this would throw the baby out with the bathwater. the peptide name and accNo would also get tossed and need to be rederived - using derivePeptideName. so now i just have it clear out the peptides residue, but leave the peptide Sequence object intact with its name. and translate() then just fills in the new residues. this is a step in the peptides maturity - its becoming less of a on the fly tag-along and evolving more towards a feature with preserved state - which is what it needs to evolve towards.
im gonna test it out but i think this also now means i can gut
derivePeptideName - woohoo! also the jdbc adapter is now setting the
peptide accession and its preserved in apollo - but its the wrong one
- at least for rice - its using the uniquename instead of the name -
need to fix that too.

jalview.gui.DrawableSequence implements AlignSequenceI which
implements SequenceI - it shouldnt - DrawableSequence should be in a
HASA with its model interface not an ISA.