fixed this bug from todo:
in most--but not every--case, if you switch data adapters, the new data
doesn't come up right.  the tiers come up collapsed to the point of
(near-)invisibility; if you then Expand all tiers, they show up.  i
didn't notice this behavior when switching between game and chado xml
adapters, but if you load something from ensj or gff and then try game or
chado xml, you'll see the problem.  the simplest way to see the problem
is to load tiny.xml and then the sample gff file--it comes up with all
the results tiny and gray.

this was my doing - the sending out of the prop scheme change event -
previously called tiers changed event - was getting triggered via the
constructor of the TypesObserver, which a new one was getting
constructed with every load - so now gui cur state tell the prop
scheme to fire one as soon as it sets it up with the controller as a
listener - at first i put the firing in setStyle but the prop scheme
may not have the controller listener yet.

(jbuilder was revealing a lot of things with tab indenting that i
reindented with spaces)