fixed this:
on jdbc read old transactions are not getting dumped! need to be -
same with loading from game! fix this right away as it will also
produce a lot of funny errors that are unclear as to why - probably
some below are due to this

there was an old data load listener in curation state
that used to work for annot change
log and never got migrated over for transaction manager clearing.