Added new Iterator and Factory patterns for Feature and FeatureFilter.  I've implemented the the overlaps= and type= filters already.
overlaps= should be able to serve as a template for contains=, identical=,
and inside=, as well as possibly the target_*= filterss (the latter of which
i don't yet fully understand, and we don't have any target data for anyway).

type= should very easily serve as a template for exacttype=.

that only leaves att=, name=, and xid=.  these should be implementable with
only a few hours work.  yay!