Rice can now bring in out of range hsps features that are part of in range hits. rice can do this because it has feature locs on hits, where fly & tigr dont. I added a configurable to chado-adapter.xml config file, <searchHitsHaveFeatLocs>true/false. Without this blast hits that get cut off are rather misleading - fly doesnt have these feat locs so cant ammend that. They intend to add those feat locs at some point, at which point can just change the configuration.
2 more tables are now being joined, feat rel & feat loc,
in the slowest of queries, the
search hits query. I fiddled with the query a bunch and got it so its
only slightly slower than it was previously surprisingly (9 seconds
went up to 12 seconds) - so thats nice.

I added AbstractChadoInstance thats the super class of the rest of
chado instances. This handles default chado instance stuff.
This may morph into DefaultChadoInstance at some
point. in fact if everything was configurable (which might be nice)
the subclasses wouldnt be necasary.

Also broke out from ChadoAdapter XmlConfigFileParser that parses the
xml config file.