Fixed bug where when EDE was doing a merge due to 2 genes orfs made to overlap (or even utrs) it was throwing an exception. that exception was causing a bunch of transactions to be dropped, which caused the rice commit to fail, and caused the merging names to be wrong. the jdbc now commits ok and the transcript merge names are coming out ok. The problem was delete transcript from annot editor clones the deleted transcript, and the ede couldnt match the clone with the real thing with ==. so it now checks with isIdentical() which picks up clones. the reason delete trans clones is that for splits and merges deleted transcripts end up with new genes, which messes up the drawables in figuring things out - the clone is not reattached of course.
ede is not reacting well to some annotation change events - it doesnt
throw exceptions and halt things but the display is coming up funny - some of
this is probably old and some might be new. the workaround is to bring
up a new EDE. one easy solution might be to wipe out the ede and
rebuild rather than tinker on ACEs, if that doesnt take too much time.

also i put in more informative error messages if ede fails to find a
feature. and the exception is supressed as well so it wont cause other
things to fail.